feat: add patient/guardian info, refactor formatters a bit
